doc: 性能测试添加测试结果注释
