+++ src/test/java/top/zhogjianhao/jmh/ToCharsTest.java
@@ -51,3 +51,41 @@ public class ToCharsTest {
     return StandardCharsets.UTF_8.decode(bb).array();
   }
 }
+
+// Benchmark                                                    Mode     Cnt     Score    Error   Units
+// ToCharsTest.charBufferAllocate                              thrpt       5    17.768 ±  1.110  ops/us
+// ToCharsTest.charBufferWrap                                  thrpt       5    21.095 ±  0.273  ops/us
+// ToCharsTest.string                                          thrpt       5    26.156 ±  4.919  ops/us
+// ToCharsTest.charBufferAllocate                               avgt       5     0.055 ±  0.003   us/op
+// ToCharsTest.charBufferWrap                                   avgt       5     0.042 ±  0.001   us/op
+// ToCharsTest.string                                           avgt       5     0.038 ±  0.005   us/op
+// ToCharsTest.charBufferAllocate                             sample  143812     0.094 ±  0.027   us/op
+// ToCharsTest.charBufferAllocate:charBufferAllocate·p0.00    sample               ≈ 0            us/op
+// ToCharsTest.charBufferAllocate:charBufferAllocate·p0.50    sample             0.100            us/op
+// ToCharsTest.charBufferAllocate:charBufferAllocate·p0.90    sample             0.100            us/op
+// ToCharsTest.charBufferAllocate:charBufferAllocate·p0.95    sample             0.100            us/op
+// ToCharsTest.charBufferAllocate:charBufferAllocate·p0.99    sample             0.200            us/op
+// ToCharsTest.charBufferAllocate:charBufferAllocate·p0.999   sample             0.800            us/op
+// ToCharsTest.charBufferAllocate:charBufferAllocate·p0.9999  sample            11.566            us/op
+// ToCharsTest.charBufferAllocate:charBufferAllocate·p1.00    sample          1167.360            us/op
+// ToCharsTest.charBufferWrap                                 sample  107152     0.068 ±  0.004   us/op
+// ToCharsTest.charBufferWrap:charBufferWrap·p0.00            sample               ≈ 0            us/op
+// ToCharsTest.charBufferWrap:charBufferWrap·p0.50            sample             0.100            us/op
+// ToCharsTest.charBufferWrap:charBufferWrap·p0.90            sample             0.100            us/op
+// ToCharsTest.charBufferWrap:charBufferWrap·p0.95            sample             0.100            us/op
+// ToCharsTest.charBufferWrap:charBufferWrap·p0.99            sample             0.100            us/op
+// ToCharsTest.charBufferWrap:charBufferWrap·p0.999           sample             0.800            us/op
+// ToCharsTest.charBufferWrap:charBufferWrap·p0.9999          sample            10.679            us/op
+// ToCharsTest.charBufferWrap:charBufferWrap·p1.00            sample           109.056            us/op
+// ToCharsTest.string                                         sample  123424     0.072 ±  0.003   us/op
+// ToCharsTest.string:string·p0.00                            sample               ≈ 0            us/op
+// ToCharsTest.string:string·p0.50                            sample             0.100            us/op
+// ToCharsTest.string:string·p0.90                            sample             0.100            us/op
+// ToCharsTest.string:string·p0.95                            sample             0.100            us/op
+// ToCharsTest.string:string·p0.99                            sample             0.100            us/op
+// ToCharsTest.string:string·p0.999                           sample             0.600            us/op
+// ToCharsTest.string:string·p0.9999                          sample            11.189            us/op
+// ToCharsTest.string:string·p1.00                            sample            49.472            us/op
+// ToCharsTest.charBufferAllocate                                 ss       5    19.740 ± 28.292   us/op
+// ToCharsTest.charBufferWrap                                     ss       5    14.240 ± 13.385   us/op
+// ToCharsTest.string                                             ss       5    13.160 ±  8.630   us/op
